+++ src/main/java/com/cyprias/chunkspawnerlimiter/Config.java
@@ -1,12 +1,9 @@
-package com.cyprias.ChunkSpawnerLimiter;
+package com.cyprias.chunkspawnerlimiter;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
 import org.bukkit.configuration.InvalidConfigurationException;
-import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.configuration.file.YamlConfiguration;
 
 public class Config {
 	public static boolean getBoolean(String property) {
