+++ src/main/java/com/cyprias/chunkspawnerlimiter/Logger.java
@@ -1,4 +1,4 @@
-package com.cyprias.ChunkSpawnerLimiter;
+package com.cyprias.chunkspawnerlimiter;
 
 import org.bukkit.ChatColor;
